+++ src/main/java/py/edu/uc/lp3/content/Serie.java
@@ -8,7 +8,7 @@ public class Serie extends Video{
 	int temporadas;
 	int [] emision; //fecha en la que salio al aire y fecha en la que se termino de emitir 
 	
-	public Serie(String nombre, String genero, int id, List<SitioReview> ratings, int episodios, int temporadas,
+	public Serie(String nombre, String genero, long id, List<SitioReview> ratings, int episodios, int temporadas,
 			int[] emision) {
 		super(nombre, genero, id, ratings);
 		this.episodios = episodios;
